refactor(emitter): split source file emit helpers
